use multi-stage and push targets

+++ service-scripts/gocd/tasks/docker/publish.sh
@@ -5,6 +5,7 @@ set -e
 : ${HOME_DIR:=/home}
 : ${SKIP_TAG_LATEST:=false}
 : ${DOCKER_TAG_REV:=true}
+: ${DOCKER_REMOVE_IMAGES:=true}
 
 if [ -z ${DOCKER_REPO} ];then
     echo ">> Publish >>> Using ${DOCKER_REPO_DEFAULT} as DOCKER_REPO name"
@@ -54,7 +55,9 @@ if [ "${SKIP_TAG_LATEST}" == "false" ];then
       IMG_FULL_NAME="${DOCKER_REG}/${DOCKER_REPO}/${IMG_NAME}:${DOCKER_TAG}@${BUILD_IMG_REPODIGEST}"
       echo ">> Full name of image: ${IMG_FULL_NAME}"
       echo ${IMG_FULL_NAME} > ${ARTIFACTS_DIR}/target/${IMG_NAME}.image_name
-      docker rmi ${DOCKER_REG}/${DOCKER_REPO}/${IMG_NAME}:${DOCKER_TAG}
+      if [[ "${DOCKER_REMOVE_IMAGES}" == "true" ]];then
+        docker rmi ${DOCKER_REG}/${DOCKER_REPO}/${IMG_NAME}:${DOCKER_TAG}
+      fi
     fi
 else
    echo ">> PUBLISH >>> Skip tagging the build as ${DOCKER_TAG}"
@@ -89,9 +92,60 @@ if [ "${DOCKER_TAG_REV}" == "true" ];then
         IMG_FULL_NAME="${DOCKER_REG}/${BUILD_REV_NAME}@${BUILD_IMG_REPODIGEST}"
         echo ">> Full name of image: ${IMG_FULL_NAME}"
         echo ${IMG_FULL_NAME} > ${ARTIFACTS_DIR}/target/${IMG_NAME}.image_name
-        docker rmi ${DOCKER_REG}/${BUILD_REV_NAME}
+        if [[ "${DOCKER_REMOVE_IMAGES}" == "true" ]];then
+          docker rmi ${DOCKER_REG}/${BUILD_REV_NAME}
+        fi
     fi
 else
    echo ">> PUBLISH >>> Skip pushing to registry, since none set"
 fi
-docker rmi ${BUILD_IMG_NAME}
+
+## Push targets
+if [[ "${DOCKER_BUILD_TARGETS}" == "false" ]];then
+  echo ">> Skip creating target images: DOCKER_BUILD_TARGETS==false"
+  exit 0
+fi
+echo ">> Creating target images: DOCKER_BUILD_TARGETS==true"
+DFILE_TARGETS=$(grep '^FROM' ${DOCKER_FILE} | awk '/^FROM.* AS /{print $NF}' |xargs |sed -e 's/ /:/g')
+for DFILE_TARGET in $(echo ${DFILE_TARGETS} |sed -e 's/:/ /g');do
+  echo ">> Targets to push: ${DFILE_TARGETS}"
+  assemble_target_img_name ${DFILE_TARGET}
+  BUILD_REV_NAME="${DOCKER_REPO}/${IMG_NAME}:${DOCKER_TAG}-${DFILE_TARGET}-rev${GO_PIPELINE_COUNTER}"
+  echo ">> Publish >>> Tag image locally with build revision: docker tag/push/rmi ${BUILD_REV_NAME}"
+  docker tag ${BUILD_IMG_NAME} ${BUILD_REV_NAME}
+  if [ "X${DOCKER_REG}" != "X" ];then
+      echo ">> Publish >>> Tag image remotely with build revision: docker tag/push/rmi ${DOCKER_REG}/${BUILD_REV_NAME}"
+      docker tag ${BUILD_IMG_NAME} ${DOCKER_REG}/${BUILD_REV_NAME}
+      set +e
+      begin=$(date +%s)
+      docker push ${DOCKER_REG}/${BUILD_REV_NAME}
+      if [[ $? -ne 0 ]];then
+          end=$(date +%s)
+          if [[ $(($end-$begin)) -lt 200 ]];then
+            echo "[!!] Push failed (<200s)"
+            exit 1
+          else
+            echo "[WW] Push failed (>200s), try once more"
+            docker push ${DOCKER_REG}/${BUILD_REV_NAME}
+            if [[ $? -ne 0 ]];then
+              echo "[!!] Push failed second time. :("
+              exit 1
+            fi
+          fi
+      fi
+      set -e
+      BUILD_IMG_REPODIGEST=$(docker inspect -f '{{(index .RepoDigests 0) }}' ${DOCKER_REG}/${BUILD_REV_NAME} |awk -F\@ '{print $2}')
+      IMG_FULL_NAME="${DOCKER_REG}/${BUILD_REV_NAME}@${BUILD_IMG_REPODIGEST}"
+      echo ">> Full name of image: ${IMG_FULL_NAME}"
+      echo ${IMG_FULL_NAME} > ${ARTIFACTS_DIR}/target/${IMG_NAME}.image_name
+      if [[ "${DOCKER_REMOVE_IMAGES}" == "true" ]];then
+        docker rmi ${DOCKER_REG}/${BUILD_REV_NAME}
+      fi
+  fi
+  if [[ "${DOCKER_REMOVE_IMAGES}" == "true" ]];then
+    ${TARGET_IMG_NAME}
+  fi
+done
+if [[ "${DOCKER_REMOVE_IMAGES}" == "true" ]];then
+  docker rmi ${BUILD_IMG_NAME}
+fi
